requirements.txt
# Configuración
python-dotenv==1.0.0

# Google Drive API
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Procesamiento de Excel/CSV
pandas==2.2.3
openpyxl==3.1.5